Move whisper from models/experimental to models/demos
