+++ models/demos/whisper/demo/demo.py
@@ -22,8 +22,8 @@ import jiwer
 
 import ttnn
 from ttnn.model_preprocessing import preprocess_model_parameters
-from models.experimental.functional_whisper.tt import ttnn_optimized_functional_whisper
-from models.experimental.functional_whisper.tt.ttnn_optimized_functional_whisper import (
+from models.demos.whisper.tt import ttnn_optimized_functional_whisper
+from models.demos.whisper.tt.ttnn_optimized_functional_whisper import (
     init_kv_cache,
     WHISPER_L1_SMALL_SIZE,
 )
